Run license vetting only on target file change
That will make some changes be found later e.g. Maven dependency
defining it's dependency version with a range and new version of this
dependency of dependency still within the range be pushed to Maven
central.
It is needed due to EF dash tool not being able to survive the load.