Fix DnD crash on Windows
Remove IdentificationTargetTransferData class (which is not used in any
other way) and getSupportedTypes as creating *TransferData without
having the native parts initialized leads to calling windows api with
nulls and thus the crash.
The implementation in isSupportedType should be enough of a guard.
In SWT's TransferData javadoc there is:
```
You should only need to become familiar with the fields in this class if
you are implementing a Transfer subclass and you are unable to subclass
the ByteArrayTransfer class.
```

If one looks at JFace implementation it doesn't override
getSupportedTypes
https://github.com/eclipse-platform/eclipse.platform.ui/blob/a3fa06583905e66550c0b4e8a3961320cfb4ca43/bundles/org.eclipse.jface/src/org/eclipse/jface/util/LocalSelectionTransfer.java#L35
most likely for the same reason.